Added token based accessing limits

+++ src/main/java/comp307/backend/booking/Service/RequestService.java
@@ -19,46 +19,41 @@ public class RequestService {
     private final UserRepository userRepository;
     private final RequestRepository requestRepository;
 
+    // Type 1 meeting
     public RequestService(UserRepository userRepository, RequestRepository requestRepository) {
         this.userRepository = userRepository;
         this.requestRepository = requestRepository;
     }
 
 
-    public Request requestBooking(String requesterEmail, String ownerEmail, LocalDateTime requestedStart, LocalDateTime requestedEnd, String message) {
-        User requester = userRepository.findById(requesterEmail).orElseThrow(() -> new RuntimeException("User " + requesterEmail + " not found."));
+    public Request requestBooking(String requesterToken, String ownerEmail, LocalDateTime requestedStart, LocalDateTime requestedEnd, String message) {
+        User requester = userRepository.findById(requesterToken).orElseThrow(() -> new RuntimeException("User " + requesterToken + " not found."));
         User owner = userRepository.findById(ownerEmail).orElseThrow(() -> new RuntimeException("User " + ownerEmail + " not found."));
 
+        if (!owner.isOwner()) return null;
         Request request = new Request(requester, owner, requestedStart, requestedEnd, message);
         requestRepository.save(request);
 
         return request;
     }
 
-    public void acceptRequest(Long requestID) {
-        Optional<Request> optionalRequest = requestRepository.findById(requestID);
-
-        if (optionalRequest.isPresent()) {
-            Request request = optionalRequest.get();
-
-            request.setStatus(true);
-            requestRepository.save(request);
-        }
-    }
-
-    public void declineRequest(Long requestID) {
+    public void setRequestState(Long requestID, String ownerToken, boolean accept) {
         Optional<Request> request = requestRepository.findById(requestID);
+        User owner = userRepository.findByaccessToken(ownerToken).orElseThrow(() -> new RuntimeException("User " + ownerToken + " not found."));
 
-        if (request.isPresent()) {
-            request.get().setStatus(false);
-            requestRepository.delete(request.get());
+        if (request.isPresent() && request.get().isPending() && request.get().getOwner().equals(owner)) {
+            if (accept) {
+                request.get().setStatus(true);
+                requestRepository.save(request.get());
+            } else {
+                request.get().setStatus(false);
+                requestRepository.delete(request.get());
+            }
         }
-
     }
-
-    public List<Request> getPendingRequests(String ownerEmail) {
+    public List<Request> getPendingRequests(String ownerToken) {
         ArrayList<Request> requests = new ArrayList<>();
-        Optional<User> owner = userRepository.findById(ownerEmail);
+        Optional<User> owner = userRepository.findByaccessToken(ownerToken);
         if (owner.isEmpty()) return requests;
 
         for (Request request : requestRepository.findByOwner(owner.get())) {
